# go-client/application_status.go
/* 
 * Fusion Creator API
 *
 * .
 *
 * OpenAPI spec version: 0.1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

package swagger

type ApplicationStatus struct {

	Id string `json:"id,omitempty"`

	Name string `json:"name"`

	Services []ServiceStatus `json:"services,omitempty"`
}