# go-client/named_link.go
/* 
 * Fusion Creator API
 *
 * .
 *
 * OpenAPI spec version: 0.1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

package swagger

type NamedLink struct {

	// The name of the service.
	Name string `json:"name"`

	// Url of the service.
	Url string `json:"url,omitempty"`
}