# go-client/service_desc.go
/* 
 * Fusion Creator API
 *
 * .
 *
 * OpenAPI spec version: 0.1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

package swagger

type ServiceDesc struct {

	Id int64 `json:"id,omitempty"`

	// The name of the service (required)
	Name string `json:"name"`

	// A SemVer 2 version
	Version string `json:"version,omitempty"`
}