# go-client/service_name.go
/* 
 * Fusion Creator API
 *
 * .
 *
 * OpenAPI spec version: 0.1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

package swagger

type ServiceName struct {

	// The name of the service.
	Name string `json:"name"`

	// Version of the service.
	Version string `json:"version,omitempty"`
}