# go-server-server/go/service.go
package server

import (
	"net/http"
)

type Service struct {

}

func ListService(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
		w.WriteHeader(http.StatusOK)
}